Disable module scanning to prevent Clang tools errors
CMake scans for C++ modules if the compiler supports it (like GCC 14)
and at least C++ 20 is used. This adds additional flags to the compile
command that Clang 19 and its tools do not know, so they throw errors.
Since we don't use modules we can just disable the scan to resolve this.