Fix using clang-tidy-cache
A while a go, I moved the option for enabling clang-tidy-cache to the
developer options in `DeveloperMode.cmake`. This was a mistake because
that file is included after the targets are defined, but to enable
clang-tidy-cache we need to modify `CMAKE_CXX_CLANG_TIDY` which only
effects targets defined after its change.